requirements.txt
streamlit==1.29.0
pandas==2.2.3
numpy==1.26.2
plotly==5.18.0
google-generativeai==0.3.1
python-dotenv==1.0.0
openpyxl==3.1.2